launcher: split codebase into multiple files
