+++ launcher/sloth.go
@@ -1,0 +1,155 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"io"
+	"os"
+	"strconv"
+)
+
+type JsonRaw = map[string]any
+type Thunk func() string
+
+type EnvVar struct {
+	Type string
+	Key  string
+	Or   Thunk
+}
+
+func NewEnvVar(raw JsonRaw) (e EnvVar) {
+	e.Type = "env"
+	e.Key = raw["key"].(string)
+	if or, ok := raw["or"]; ok {
+		e.Or = func() string {
+			return valToString(or)
+		}
+	}
+	return
+}
+
+func (e EnvVar) String() string {
+	r, _ := os.LookupEnv(e.Key)
+	if r != "" {
+		return r
+	}
+	if e.Or != nil {
+		return e.Or()
+	}
+	panic(fmt.Sprintf("environment variable '%s' not set", e.Key))
+}
+
+type Concat struct {
+	Type string
+	A    string
+	B    string
+}
+
+func NewConcat(raw JsonRaw) (c Concat) {
+	c.Type = "concat"
+	c.A = valToString(raw["a"])
+	c.B = valToString(raw["b"])
+	return
+}
+
+func (c Concat) String() string {
+	return c.A + c.B
+}
+
+type InstanceId struct {
+	Type string
+	Id   string
+}
+
+func NewInstanceId(raw JsonRaw) (i InstanceId) {
+	i.Type = "instanceId"
+	i.Id = instanceId()
+	return
+}
+
+func (i InstanceId) String() string {
+	return i.Id
+}
+
+type Uid struct {
+	Type string
+}
+
+func NewUid(raw JsonRaw) (u Uid) {
+	u.Type = "uid"
+	return
+}
+
+func (u Uid) String() string {
+	return strconv.Itoa(os.Getuid())
+}
+
+type Gid struct {
+	Type string
+}
+
+func NewGid(raw JsonRaw) (u Gid) {
+	u.Type = "gid"
+	return
+}
+
+func (u Gid) String() string {
+	return strconv.Itoa(os.Getgid())
+}
+
+type Mkdir struct {
+	Type string
+	Dir  string
+}
+
+func NewMkdir(raw JsonRaw) (m Mkdir) {
+	m.Type = "mkdir"
+	m.Dir = valToString(raw["dir"])
+	return
+}
+
+func (m Mkdir) String() string {
+	err := os.MkdirAll(m.Dir, 0700)
+	if err != nil {
+		fmt.Println(err)
+	}
+	return m.Dir
+}
+
+func valToString(item any) (ret string) {
+	ret, ok := item.(string)
+	if ok {
+		return
+	} else {
+		raw, _ := item.(JsonRaw)
+		switch raw["type"] {
+		case "env":
+			ret = NewEnvVar(raw).String()
+		case "concat":
+			ret = NewConcat(raw).String()
+		case "instanceId":
+			ret = NewInstanceId(raw).String()
+		case "uid":
+			ret = NewUid(raw).String()
+		case "gid":
+			ret = NewGid(raw).String()
+		case "mkdir":
+			ret = NewMkdir(raw).String()
+		default:
+			panic("Unknown type: \"" + raw["type"].(string) + "\"")
+		}
+		return
+	}
+}
+
+func readJsonArgs(filename string) (args []string) {
+	file, _ := os.Open(filename)
+	bytes, _ := io.ReadAll(file)
+	var argsRaw []any
+
+	json.Unmarshal(bytes, &argsRaw)
+	for _, item := range argsRaw {
+		args = append(args, valToString(item))
+	}
+	return
+}
